+++ R/y.R
@@ -18,6 +18,7 @@
 #' points to a reference.
 #' @param pkg The package containing the reference. The package must be loaded
 #' and attached. The default `pkg = "yzy"` searches in the home package.
+#' @param verbose Set to `TRUE` to turn on warnings
 #' @param dec A scalar value indicating the number of decimals used to round the
 #' value. The default is 3 decimals.
 #' @param tail_adjust Logical. If \code{TRUE} the procedure applies the WHO
@@ -37,7 +38,8 @@
 #' )
 #' y <- y(z, x, refcode)
 #' @export
-y <- function(z, x, refcode, pkg = "yzy", dec = 3L, tail_adjust = FALSE, ...) {
+y <- function(z, x, refcode, pkg = "yzy", verbose = FALSE,
+              dec = 3L, tail_adjust = FALSE, ...) {
   if (length(z) != length(x) || length(z) != length(refcode)) {
     message("y(): Non-conformable arguments")
     return(rep(NA_real_, length(z)))
@@ -52,6 +54,7 @@ y <- function(z, x, refcode, pkg = "yzy", dec = 3L, tail_adjust = FALSE, ...) {
       z = .data$z,
       x = .data$x,
       refcode = first(.data$refcode),
+      verbose = verbose,
       pkg = pkg,
       tail_adjust = tail_adjust
     )) %>%
@@ -60,8 +63,8 @@ y <- function(z, x, refcode, pkg = "yzy", dec = 3L, tail_adjust = FALSE, ...) {
     round(digits = dec)
 }
 
-y_grp <- function(z, x, refcode, pkg, tail_adjust = FALSE) {
-  r <- load_reference(refcode = refcode, pkg = pkg)
+y_grp <- function(z, x, refcode, pkg, verbose, tail_adjust = FALSE) {
+  r <- load_reference(refcode = refcode, pkg = pkg, verbose = verbose)
 
   # do not process in absence of study attribute
   study <- attr(r, "study")